requirements.txt
Markdown==2.6.8
python-frontmatter==0.4.2
PyYAML==3.12
six==1.10.0
